#47 Move config to one file. Write unit tests for builder
