+++ source/utils/config/common.py
@@ -13,3 +13,10 @@ class BaseSchema(Generic[T], ABC):
     def build(self, data: Dict) -> T:
         raise NotImplementedError()
 
+class InvalidConfigException(Exception):
+    def __init__(self, msg):
+        super().__init__(msg)
+
+class MissingParsingLibrary(Exception):
+    def __init__(self, library, parserName) -> None:
+        super().__init__(f"Missing {library}. Required for {parserName}")
