--- source/utils/config/exceptions.py
@@ -1,4 +1,0 @@
-class InvalidConfigException(Exception):
-    def __init__(self, msg):
-        super().__init__(msg)
-
