--- tests/testConfig.py
@@ -1,3 +1,0 @@
-import unittest
-
-
